+++ packages/dnr-rulesets/CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [3.1.1] - 2025-07-02
 
-## Added
+### Added
 
 - New `exclude-unsafe-rules` CLI command to exclude unsafe rules from rulesets
   and save them to the metadata of rulesets for use "skip review" feature in CWS.
@@ -15,6 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   and automatically rebuild DNR.
 - New option `--latest-filters` for `load` command to load fresh text filters
   instead of DNR rulesets and save them to the specified directory.
+
+### Updated
+
+- Updated [@adguard/tsurlfilter] to `v3.4.1`.
+
+[3.1.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/dnr-rulesets-v3.1.1
 
 ## [3.1.0] - 2025-06-06
 
@@ -25,6 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated [@adguard/logger] to `v2.0.0`.
+- Updated [@adguard/tsurlfilter] to `v3.4.0`.
+
+[3.1.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/dnr-rulesets-v3.1.0
 
 ## [3.0.0-alpha.12] - 2025-05-28